+++ app/src/main/java/woowacourse/movie/MoviesAdapter.kt
@@ -8,6 +8,7 @@ import android.widget.BaseAdapter
 import android.widget.Button
 import android.widget.ImageView
 import android.widget.TextView
+import woowacourse.movie.domain.Movie
 
 class MoviesAdapter(
 //    private val context: Context,
