+++ app/src/main/java/woowacourse/movie/domain/Movie.kt
@@ -1,7 +1,8 @@
-package woowacourse.movie
+package woowacourse.movie.domain
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
+import woowacourse.movie.R
 import java.time.LocalDate
 
 @Parcelize
