+++ app/src/main/java/woowacourse/movie/domain/TicketInfo.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie
+package woowacourse.movie.domain
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
