+++ app/src/main/java/woowacourse/movie/domain/Time.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie
+package woowacourse.movie.domain
 
 import java.time.LocalTime
